requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
gunicorn
flake8
python-dotenv